Backend types to shared folder
Moved types that are used in both frontend and backend from
customTypes and jiraTypes.
Deleted duplicate types from frontend.
Updated imports.
