+++ frontend/src/components/RoadmapOverview.tsx
@@ -4,7 +4,8 @@ import { shallowEqual, useSelector } from 'react-redux';
 import classNames from 'classnames';
 import MonetizationOnIcon from '@material-ui/icons/MonetizationOn';
 import { chosenRoadmapSelector } from '../redux/roadmaps/selectors';
-import { Roadmap, TaskRatingDimension } from '../redux/roadmaps/types';
+import { Roadmap } from '../redux/roadmaps/types';
+import { TaskRatingDimension } from '../../../shared/types/customTypes';
 import { RootState } from '../redux/types';
 import { roadmapsVersionsSelector } from '../redux/versions/selectors';
 import { Version } from '../redux/versions/types';
