+++ frontend/src/components/TableCustomerRow.tsx
@@ -9,7 +9,8 @@ import { ModalTypes } from '../redux/modals/types';
 import { Customer } from '../redux/roadmaps/types';
 import { RootState } from '../redux/types';
 import { userInfoSelector } from '../redux/user/selectors';
-import { UserInfo, UserType } from '../redux/user/types';
+import { UserInfo } from '../redux/user/types';
+import { UserType } from '../../../shared/types/customTypes';
 import css from './TableCustomerRow.module.scss';
 
 const classes = classNames.bind(css);
