+++ frontend/src/components/TableTaskRow.tsx
@@ -12,7 +12,8 @@ import { roadmapsActions } from '../redux/roadmaps/index';
 import { Task } from '../redux/roadmaps/types';
 import { RootState } from '../redux/types';
 import { userInfoSelector } from '../redux/user/selectors';
-import { UserInfo, UserType } from '../redux/user/types';
+import { UserInfo } from '../redux/user/types';
+import { UserType } from '../../../shared/types/customTypes';
 import { DeleteButton } from './forms/DeleteButton';
 import { EditButton } from './forms/EditButton';
 import { InfoButton } from './forms/InfoButton';
