+++ frontend/src/components/TableTeamMemberRow.tsx
@@ -8,7 +8,8 @@ import { DeleteButton } from './forms/DeleteButton';
 import { EditButton } from './forms/EditButton';
 import { BusinessValueFilled } from './RatingIcons';
 import { RoadmapUser } from '../redux/roadmaps/types';
-import { RoleType, UserType, UserInfo } from '../redux/user/types';
+import { UserInfo } from '../redux/user/types';
+import { RoleType, UserType } from '../../../shared/types/customTypes';
 import css from './TableTeamMemberRow.module.scss';
 import { RootState } from '../redux/types';
 import { userInfoSelector } from '../redux/user/selectors';
