+++ frontend/src/components/TaskHeatmap.tsx
@@ -3,7 +3,8 @@ import React from 'react';
 import classNames from 'classnames';
 import { shallowEqual, useSelector } from 'react-redux';
 import { chosenRoadmapSelector } from '../redux/roadmaps/selectors';
-import { Roadmap, TaskRatingDimension } from '../redux/roadmaps/types';
+import { Roadmap } from '../redux/roadmaps/types';
+import { TaskRatingDimension } from '../../../shared/types/customTypes';
 import { RootState } from '../redux/types';
 import { calcTaskAverageRating } from '../utils/TaskUtils';
 import css from './TaskHeatmap.module.scss';
