+++ frontend/src/components/TaskValueCreatedVisualization.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { shallowEqual, useSelector } from 'react-redux';
 import { Cell, Legend, Pie, PieChart, Tooltip } from 'recharts';
 import { allCustomersSelector } from '../redux/roadmaps/selectors';
-import { Customer, TaskRatingDimension } from '../redux/roadmaps/types';
+import { Customer } from '../redux/roadmaps/types';
+import { TaskRatingDimension } from '../../../shared/types/customTypes';
 import { RootState } from '../redux/types';
 import { Version } from '../redux/versions/types';
 import { calcTaskValueSum } from '../utils/TaskUtils';
