+++ frontend/src/components/UserInfoCard.tsx
@@ -3,7 +3,8 @@ import { Trans } from 'react-i18next';
 import { shallowEqual, useSelector } from 'react-redux';
 import { RootState } from '../redux/types';
 import { userInfoSelector } from '../redux/user/selectors';
-import { UserInfo, UserType } from '../redux/user/types';
+import { UserInfo } from '../redux/user/types';
+import { UserType } from '../../../shared/types/customTypes';
 
 export const UserInfoCard = () => {
   const userInfo = useSelector<RootState, UserInfo | undefined>(
