+++ frontend/tsconfig.json
@@ -21,5 +21,6 @@
   },
   "include": [
     "src"
-  ]
+  ],
+  "extends": "./tsconfig.paths.json"
 }
