+++ server/src/api/customer/customer.routes.ts
@@ -1,5 +1,6 @@
 import { requirePermission } from './../../utils/checkPermissions';
-import { IKoaState, Permission } from './../../types/customTypes';
+import { IKoaState } from './../../types/customTypes';
+import { Permission } from '../../../../shared/types/customTypes';
 import KoaRouter from '@koa/router';
 import {
   getCustomers,
