+++ server/src/seeds/testdata.ts
@@ -4,7 +4,7 @@ import Roadmap from '../api/roadmaps/roadmaps.model';
 import User from '../api/users/users.model';
 import { Role } from '../api/roles/roles.model';
 import Version from '../api/versions/versions.model';
-import { UserType, RoleType } from '../types/customTypes';
+import { UserType, RoleType } from '../../../shared/types/customTypes';
 import Customer from '../api/customer/customer.model';
 
 export async function seed(knex: Knex): Promise<any> {
