+++ server/src/types/jiraTypes.ts
@@ -1,9 +1,9 @@
 export interface JiraOAuthURLResponse {
-    url: string;
-    token: string;
-    token_secret: string;
+  url: string;
+  token: string;
+  token_secret: string;
 }
 
 export interface JiraOAuthRequestTokenResponse {
-    token: string;
+  token: string;
 }
