+++ server/tests/roadmaps.test.ts
@@ -3,7 +3,7 @@ import chaiHttp from 'chai-http';
 import Roadmap from '../src/api/roadmaps/roadmaps.model';
 import User from '../src/api/users/users.model';
 import { Role } from '../src/api/roles/roles.model';
-import { Permission, RoleType } from '../src/types/customTypes';
+import { Permission, RoleType } from '../../shared/types/customTypes';
 import { loggedInAgent } from './setuptests';
 chai.use(chaiHttp);
